refactor: move constants
