+++ src/App.tsx
@@ -4,8 +4,8 @@ import {BrowserRouter} from 'react-router-dom';
 import {Container} from 'styles/index';
 import Header from 'components/UI/Header/Header';
 import AppRouter from 'router/AppRouter';
-import {LocalStorageKeys} from 'constants/localStorage';
-import {getDataBag, partInitBag} from 'constants/mock';
+import {LocalStorageKeys} from 'utils/localStorage';
+import {partInitBag} from 'mock/bag.mock';
 import {getItem, setItem} from 'utils/localStorage';
 import {useActions} from 'hooks/useActions';
 import {IBagState} from 'store/slices/bag.slice';
@@ -13,6 +13,7 @@ import {useGetAssetsQuery, useLazyGetAssetsQuery} from 'api/endPoints';
 import {Loader} from 'components/UI/Loader/Loader.styled';
 import {useTypedSelector} from 'hooks/useTypedSelector';
 import {IAsset} from 'types/api';
+import {getDataBag} from 'utils/bag';
 
 const App: FC = () => {
   const ids: string = useTypedSelector(state => state.bag.ids).join();
