+++ src/api/constants.ts
@@ -1,0 +1,6 @@
+export const apiTags = {
+  bag: 'Bag',
+}
+
+export const limit: number = 10;
+export const totalCurrency: number = 100;
