--- src/constants/localStorage.ts
@@ -1,4 +1,0 @@
-export enum LocalStorageKeys {
-  BAG = 'bag',
-  MAIN_PAG_ITEM = 'mainPagItem'
-}
