--- src/constants/slices.ts
@@ -1,4 +1,0 @@
-export enum SliceNames {
-  COMMON = 'common',
-  BAG = 'bag'
-}
