--- src/constants/ui.ts
@@ -1,1 +1,0 @@
-export const maxNumber: number = 1000000000;
