+++ src/utils/bag.ts
@@ -1,4 +1,6 @@
-import {ICurrency, IHistory} from 'types/bag';
+import {IBagState} from 'store/slices/bag.slice';
+import {IAsset} from 'types/api';
+import {ICurrency, ICurrencyPartInitBag, IHistory, IPartInitBag} from 'types/bag';
 import {IListItem} from 'types/ui';
 import {roundNumber} from 'utils/common';
 
@@ -43,4 +45,34 @@ export function getTotalNumber(history: IHistory[], priceUsd?: number): number {
 
 export function isSpent(history: IHistory[]): boolean {
   return getTotalNumber(history) === 0;
+}
+
+function getHistoryById(currency: ICurrencyPartInitBag[] | ICurrency[], id: string): IHistory[] {
+  return currency.find(currencyOne => currencyOne.id === id)?.history || [];
+}
+
+export function getDataBag(lastBag: IPartInitBag | IBagState, assets: IAsset[]): IBagState {
+  const idInitCurrency: string[] = lastBag.ids;
+  
+  const assetsInitCurrency: IAsset[] = [];
+  assets.forEach(asset => {
+    if(idInitCurrency.includes(asset.id)) {
+      assetsInitCurrency.push(asset);
+    }
+  });
+
+  const currency: ICurrency[] = assetsInitCurrency.map(asset => ({
+    id: asset.id,
+    name: asset.name,
+    symbol: asset.symbol,
+    priceUsd: Number(asset.priceUsd),
+    history: getHistoryById(lastBag.currency, asset.id)
+  }));
+
+  const initBag: IBagState = {
+    currency,
+    ids: idInitCurrency
+  }
+
+  return initBag;
 }
